refactor into two separate crates
