+++ Cargo.toml
@@ -60,7 +60,8 @@ members = [
   "token/transfer-hook/cli",
   "token/transfer-hook/example",
   "token/transfer-hook/interface",
-  "token/confidential-transfer/proofs",
+  "token/confidential-transfer/proof-extraction",
+  "token/confidential-transfer/proof-generation",
   "token/client",
   "utils/cgen",
   "utils/test-client",
