+++ token/confidential-transfer/proof-extraction/src/errors.rs
@@ -1,0 +1,11 @@
+use thiserror::Error;
+
+#[derive(Error, Clone, Debug, Eq, PartialEq)]
+pub enum TokenProofExtractionError {
+    #[error("ElGamal pubkey mismatch")]
+    ElGamalPubkeyMismatch,
+    #[error("Pedersen commitment mismatch")]
+    PedersenCommitmentMismatch,
+    #[error("Range proof length mismatch")]
+    RangeProofLengthMismatch,
+}
