+++ token/confidential-transfer/proof-generation/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "spl-token-confidential-transfer-proof-generation"
+version = "0.1.0"
+description = "Solana Program Library Confidential Transfer Proofs"
+authors = ["Solana Labs Maintainers <maintainers@solanalabs.com>"]
+repository = "https://github.com/solana-labs/solana-program-library"
+license = "Apache-2.0"
+edition = "2021"
+
+[dependencies]
+curve25519-dalek = "3.2.1"
+solana-zk-sdk = "2.0.0"
+thiserror = "1.0.61"
+
+[dev-dependencies]
+
+[lib]
+crate-type = ["cdylib", "lib"]
