+++ token/confidential-transfer/proof-generation/src/errors.rs
@@ -1,0 +1,11 @@
+use {solana_zk_sdk::zk_elgamal_proof_program::errors::ProofGenerationError, thiserror::Error};
+
+#[derive(Error, Clone, Debug, Eq, PartialEq)]
+pub enum TokenProofGenerationError {
+    #[error("inner proof generation failed")]
+    ProofGeneration(#[from] ProofGenerationError),
+    #[error("not enough funds in account")]
+    NotEnoughFunds,
+    #[error("illegal amount bit length")]
+    IllegalAmountBitLength,
+}
